{"schema":"decodex/commit/1","summary":"Split admin and dreaming policy modules","authority":"manual"}
